add gejala

+++ blueprint_profile.py
@@ -5,6 +5,8 @@ import database as db
 import pandas as pd
 from flask_paginate import Pagination, get_page_parameter
 
+PER_PAGES = 5
+
 bp_profile = Blueprint("bp_profile",__name__)
 
 @bp_profile.route("/user-profile",methods=["POST","GET"])
@@ -31,6 +33,10 @@ def user_profile():
 @bp_profile.route('/update-users', methods=['POST',"GET"])
 def update_users():
 
+    if "google_login" not in session:
+        restricted_message = ""
+        return render_template('login.html')
+
     hasil = request.form.to_dict(flat=False)
     string = ""
     for i in hasil.items():
@@ -51,6 +57,10 @@ def update_users():
 @bp_profile.route('/list_user', methods=['POST',"GET"])
 def list_users():
 
+    if "google_login" not in session:
+        restricted_message = ""
+        return render_template('login.html')
+
     query = '''
     select *
     from users
@@ -64,8 +74,11 @@ def list_users():
 @bp_profile.route('/delete_list_user', methods=['POST',"GET"])
 def delete_list_users():
 
+    if "google_login" not in session:
+        restricted_message = ""
+        return render_template('login.html')
+
     userId = request.args.get("userId")
-    print(userId)
 
     query = """
     delete from users
@@ -78,20 +91,9 @@ def delete_list_users():
 
 @bp_profile.route('/list_penyakit', methods=['POST',"GET"])
 def list_penyakit():
-
-    # page = request.args.get('page', 1, type=int)
-
-    # query = '''
-    # select *
-    # from penyakit
-    # '''
-    # all_penyakit = db.df_query(query)
-    # list_penyakit = all_penyakit.values.tolist()
-    # list_penyakit = list_penyakit.paginate(page=page, per_page=5)
-
-    # print(list_penyakit)
-
-
+    if "google_login" not in session:
+        restricted_message = ""
+        return render_template('login.html')
 
     search = False
     q = request.args.get('q')
@@ -107,17 +109,51 @@ def list_penyakit():
     all_penyakit = db.df_query(query)
     list_penyakit = all_penyakit.values.tolist()
 
-    i=(page-1)*5
+    i=(page-1)*PER_PAGES
     penyakit_page = list_penyakit[i:i+5]
+    pagination = Pagination(page=page,per_page=PER_PAGES, total=len(list_penyakit), search=search, record_name='List')
+
+    return render_template('content_admin_list_penyakit.html',penyakit = penyakit_page, pagination=pagination,css_framework='bootstrap4')
+
+@bp_profile.route('/delete_penyakit', methods=['POST',"GET"])
+def delete_penyakit():
+
+    if "google_login" not in session:
+        restricted_message = ""
+        return render_template('login.html')
+
+    penyakitId = request.args.get("penyakitId")
+    query = """
+    delete from penyakit
+    where penyakitId like '%s'
+    """%(penyakitId)
+
+    db.execute_query(query)
+
+    query = """
+    delete from relasi 
+    where penyakitId like '%s'
+    """%(penyakitId)
+
+    db.execute_query(query)
 
 
-    # pagination = Pagination(page=page, total=len(list_penyakit), record_name='list_penyakit')
-    pagination = Pagination(page=page,per_page=5, total=len(list_penyakit), search=search, record_name='List')
+    return redirect("/list_penyakit")
 
-    return render_template('content_admin_list_penyakit.html',penyakit = penyakit_page, pagination=pagination,show_single_page=True)
+@bp_profile.route('/add_penyakit', methods=['POST',"GET"])
+def add_penyakit():
+
+    return 'oke'
+
 
 @bp_profile.route('/list_gejala', methods=['POST',"GET"])
 def list_gejala():
+
+    if "google_login" not in session:
+        restricted_message = ""
+        return render_template('login.html')
+
+    page = request.args.get(get_page_parameter(), type=int, default=1)
 
     query = '''
     select *
@@ -126,14 +162,49 @@ def list_gejala():
     all_gejala = db.df_query(query)
     list_gejala = all_gejala.values.tolist()
 
-    print(list_gejala)
+    i=(page-1)*PER_PAGES
+    gejala_page = list_gejala[i:i+5]
+    pagination = Pagination(page=page,per_page=PER_PAGES, total=len(list_gejala), record_name='List')
 
-    return render_template('content_admin_list_gejala.html',gejala = list_gejala)
+
+
+    return render_template('content_admin_list_gejala.html',gejala = gejala_page,pagination = pagination)
+
+@bp_profile.route('/delete_gejala', methods=['POST',"GET"])
+def delete_gejala():
+
+    if "google_login" not in session:
+        restricted_message = ""
+        return render_template('login.html')
+
+    gejalaId = request.args.get("gejalaId")
+    query = """
+    delete from gejala
+    where gejalaId like '%s'
+    """%(gejalaId)
+    db.execute_query(query)
+
+    query = """
+    delete from relasi
+    where gejalaId like '%s'
+    """%(gejalaId)
+    db.execute_query(query)
+    return redirect("/list_gejala")
+
+
+@bp_profile.route('/add_gejala', methods=['POST',"GET"])
+def add_gejala():
+    if "google_login" not in session:
+        restricted_message = ""
+        return render_template('login.html')
+
+    return render_template('content_admin_add_gejala.html')
 
 @bp_profile.route('/list_drug', methods=['POST',"GET"])
 def list_drug():
-
-
+    if "google_login" not in session:
+        restricted_message = ""
+        return render_template('login.html')
 
     return render_template('content_admin_list_drug.html')
